feat: Refactored map component to use Google Maps and removed Leaflet and React-Leaflet libraries
