+++ package.json
@@ -10,17 +10,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@googlemaps/markerclusterer": "^2.5.3",
     "@nivo/core": "^0.85.1",
     "@nivo/pie": "^0.85.1",
     "@vis.gl/react-google-maps": "^0.8.2",
     "bootstrap": "^5.3.3",
     "bootstrap-icons": "^1.11.3",
     "ldrs": "^1.0.1",
-    "leaflet": "^1.9.4",
     "randomcolor": "^0.6.2",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-leaflet": "^4.2.1"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.56",
