+++ src/components/DinosaurNews.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useContext } from "react";
+import React, { useState, useContext } from "react";
 import { AppContext } from "../context/Context";
 
 export default function DinosaurNews() {
